AMLOnEdge/grpc-sample/requirements.txt
pandas
h5py==2.10.0
azureml-core==1.30.0
astor==0.8.1
grpcio==1.38.0
protobuf==3.17.3
scikit-learn==0.22